chore(server/notifications): code cleanup
